# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Jan Setu API
  version: 0.2.0
  description: 'Jan Setu is a civic complaint platform: citizens file complaints over a React web portal
    or a WhatsApp bot, both backed by this shared FastAPI service. This spec is generated directly from
    the live FastAPI route definitions (`app.openapi()`), then annotated with the user story each operation
    implements (`x-user-stories`, referencing the JAN-SETU smart user stories US-01..US-10) and with the
    non-2xx responses each handler actually raises.


    Two authentication schemes are in play:

    - **Citizen bearer token** (`Authorization: Bearer <access_token>`) issued by `/auth/*`, required
    by every `/api/grievances/*` route.

    - **Official bearer token** (`Authorization: Bearer <access_token>`) issued by `/api/official/auth/verify`,
    required by every other `/api/official/*` route and audience-scoped separately from citizen tokens.

    - **X-API-Key header**, required by the internal `/v1/contacts` and `/v1/messages*` routes.


    All endpoints also return `422` with FastAPI''s standard validation-error body on malformed request
    bodies/query params; that response is omitted per-operation below to avoid repeating it, but applies
    everywhere a request body or typed query parameter is present.'
paths:
...
  /v1/contacts:
    get:
      tags:
      - contacts
      - internal
      summary: Get Contacts
      operationId: get_contacts_v1_contacts_get
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      - name: x-api-key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ContactRead'
                title: Response Get Contacts V1 Contacts Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Missing or invalid X-API-Key header.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '503':
          description: API_KEY is not configured outside development.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
      security:
      - apiKeyAuth: []
      description: Lists stored WhatsApp contacts for internal debugging/ops. Requires the X-API-Key header.
  /v1/messages:
    get:
      tags:
      - messages
      - internal
      summary: Get Messages
      operationId: get_messages_v1_messages_get
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 500
          minimum: 1
          default: 100
          title: Limit
      - name: wa_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Wa Id
      - name: x-api-key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MessageRead'
                title: Response Get Messages V1 Messages Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Missing or invalid X-API-Key header.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '503':
          description: API_KEY is not configured outside development.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
      security:
      - apiKeyAuth: []
      description: Lists stored inbound/outbound WhatsApp messages, optionally filtered by wa_id. Requires
        the X-API-Key header.
  /v1/messages/text:
    post:
      tags:
      - messages
      - internal
      summary: Send Text Message
      operationId: send_text_message_v1_messages_text_post
      parameters:
      - name: x-api-key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendTextRequest'
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SendTextResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Missing or invalid X-API-Key header.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '503':
          description: API_KEY not configured outside development, or the WhatsApp client is unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '502':
          description: WhatsApp Cloud API returned an error status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
      security:
      - apiKeyAuth: []
      description: Sends a free-form WhatsApp text message via the Cloud API and stores the outbound record.
        Requires the X-API-Key header; only deliverable inside Meta's 24-hour customer-service window
        unless a template is used.
  /whatsapp/webhook:
    get:
      tags:
      - whatsapp
      summary: Verify Webhook
      operationId: verify_webhook_whatsapp_webhook_get
      parameters:
      - name: hub.mode
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Hub.Mode
      - name: hub.verify_token
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Hub.Verify Token
      - name: hub.challenge
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Hub.Challenge
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '403':
          description: hub.verify_token does not match the configured value.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
      x-user-stories:
      - US-01
      - US-02
      - US-03
      - US-09
      - US-10
      description: 'Meta''s webhook subscription handshake: echoes back hub.challenge if hub.verify_token
        matches the configured WHATSAPP_VERIFY_TOKEN.'
    post:
      tags:
      - whatsapp
      summary: Receive Webhook
      operationId: receive_webhook_whatsapp_webhook_post
      parameters:
      - name: x-hub-signature-256
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Hub-Signature-256
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Receive Webhook Whatsapp Webhook Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: X-Hub-Signature-256 does not match the HMAC of the raw body.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '400':
          description: Body is not valid JSON or is not a JSON object.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '503':
          description: WHATSAPP_APP_SECRET is not configured outside development.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
      x-user-stories:
      - US-01
      - US-02
      - US-03
      - US-09
      - US-10
      security:
      - metaWebhookSignature: []
      description: Entry point for all inbound WhatsApp traffic. Verifies the X-Hub-Signature-256 HMAC,
        persists the raw event, and hands processing to a background task so a slow or failing pipeline
        never blocks Meta's delivery or triggers a retry.
...
components:
  schemas:
...
    ContactRead:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        wa_id:
          type: string
          title: Wa Id
        profile_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Profile Name
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - id
      - wa_id
      - profile_name
      - created_at
      - updated_at
      title: ContactRead
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    MessageRead:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        contact_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Contact Id
        meta_message_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Meta Message Id
        direction:
          type: string
          title: Direction
        message_type:
          type: string
          title: Message Type
        text_body:
          anyOf:
          - type: string
          - type: 'null'
          title: Text Body
        media_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Media Id
        media_mime_type:
          anyOf:
          - type: string
          - type: 'null'
          title: Media Mime Type
        location_latitude:
          anyOf:
          - type: number
          - type: 'null'
          title: Location Latitude
        location_longitude:
          anyOf:
          - type: number
          - type: 'null'
          title: Location Longitude
        location_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Location Name
        location_address:
          anyOf:
          - type: string
          - type: 'null'
          title: Location Address
        location_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Location Url
        raw_payload:
          additionalProperties: true
          type: object
          title: Raw Payload
        received_at:
          type: string
          format: date-time
          title: Received At
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - id
      - contact_id
      - meta_message_id
      - direction
      - message_type
      - text_body
      - media_id
      - media_mime_type
      - location_latitude
      - location_longitude
      - location_name
      - location_address
      - location_url
      - raw_payload
      - received_at
      - created_at
      - updated_at
      title: MessageRead
...
    SendTextRequest:
      properties:
        to:
          type: string
          maxLength: 32
          minLength: 5
          title: To
        body:
          type: string
          maxLength: 4096
          minLength: 1
          title: Body
      type: object
      required:
      - to
      - body
      title: SendTextRequest
    SendTextResponse:
      properties:
        provider_response:
          additionalProperties: true
          type: object
          title: Provider Response
        stored_message:
          $ref: '#/components/schemas/MessageRead'
      type: object
      required:
      - provider_response
      - stored_message
      title: SendTextResponse
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    ErrorDetail:
      type: object
      properties:
        detail:
          description: Human-readable error message (or structured object for upstream provider errors).
      required:
      - detail
...
  securitySchemes:
    citizenBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'Access token from /auth/status or /auth/refresh, sent as `Authorization: Bearer <token>`.'
    officialBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token from /api/official/auth/verify, audience-scoped separately from citizen
        tokens.
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Static internal API key for /v1/contacts and /v1/messages*, set via the API_KEY setting.
    metaWebhookSignature:
      type: apiKey
      in: header
      name: X-Hub-Signature-256
      description: HMAC-SHA256 of the raw request body, signed by Meta with WHATSAPP_APP_SECRET. Not caller-supplied.
servers:
- url: http://localhost:8000
  description: Local development
...